requirements.txt
streamlit==1.38.0
langchain==0.3.0
langchain-community==0.3.0
langchain-google-genai==2.0.0
langchain-huggingface==0.1.0
faiss-cpu==1.9.0
python-dotenv==1.0.1
pickle5==0.0.11